% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggreg_risk.R
\name{aggregRisk}
\alias{aggregRisk}
\title{Aggregated Modifiable Risk Stratification}
\usage{
aggregRisk(data, riskvars = c("Dyslipidemia", "HighBP", "ImpairBG",
  "OwtObe"), grades = list(high = c(2, 3), median = 1, low = 0))
}
\arguments{
\item{data}{Data frame fo raw data set.}

\item{riskvars}{Vector of variable names. Default \code{c("Dyslipidemia", "HighBP",
"ImpairBG", "OwtObe")}}

\item{grades}{A list defining values representing high, median and low risks.
Default \code{list(high=c(2,3), median=1, low=0))}}
}
\value{
A data frame.
}
\description{
Stratify modifiable risks and aggregate the results.
}
\examples{
\dontrun{
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
