% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{asesrpt}
\alias{asesrpt}
\alias{asesrpt-package}
\title{asesrpt: An R Reporting Toolkit for ASES}
\description{
An analytics toolkit comprising of a series of work functions. The toolkit has
a lot of featured funcionalities specifically designed for ASES.
}
\details{
The toolkit comprises of several types of functions:
}
\seealso{
\pkg{\link{recharts}}
}
\author{
\strong{Maintainer}: Yiying Wang, \email{wangy@aetna.com}
}
\keyword{internal}
