% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bu_trend.R
\name{figTrendBU}
\alias{figTrendBU}
\title{BU Trend Chart in Auto Reporting}
\usage{
figTrendBU(data, imput = NULL, riskvar, tblBU, librisk, sort = list(col
  = 1, desc = FALSE), digits, lang = c("cn", "en"),
  na.string = getOption("na.string"), altfont = "Microsoft YaHei",
  dynamic = FALSE, ...)
}
\arguments{
\item{data}{Raw data frame.}

\item{imput}{Summarized data frame.}

\item{riskvar}{Variable name of the risk.}

\item{tblBU}{Data frame of the BU pool.}

\item{librisk}{Data frame of the risk pool.}

\item{sort}{A list. Default \code{list(col=1, desc=FALSE)}, meaning sort the
results ascendingly by Col1.}

\item{digits}{Number of digits of the results. Default 1.}

\item{lang}{'cn' or 'en'.}

\item{na.string}{A named vector \code{c(cn=..., en=...)}. How to present NAs.
Default \code{getOption("na.string")}. You can define yours.}

\item{altfont}{Alternative font name. Default 'Microsoft YaHei'.}

\item{dynamic}{Interactive JavaScript charts/tables or not. If TRUE, call
\pkg{recharts}::\code{\link{echartR}} to make the object. Default FALSE.}

\item{...}{}
}
\value{
A \code{ggplot2} object or \code{echarts} object.
}
\description{
Only used for automated reporting. Get BU trend distribution.
}
\examples{

}
\seealso{
\code{\link{ggplot2}}  \code{\link{echartR}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
