% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_wrapper.R
\name{getStrInterpret}
\alias{getStrInterpret}
\title{Structured Interpretation}
\usage{
getStrInterpret(librisk, riskvar, risk.level = 2,
  title.headlevel = risk.level + 1, lang = c("cn", "en"),
  suppl = list(posttitle = NULL, pretabBU = NULL, posttabBU = NULL,
  prefigAge = NULL, postfigAge = NULL, pretabAge = NULL, posttabAge = NULL,
  prefigBU = NULL, postfigBU = NULL, pretabCity = NULL, posttabCity = NULL,
  prefigCity = NULL, postfigCity = NULL, pretabTrendBU = NULL,
  posttabTrendBU = NULL, prefigTrendBU = NULL, postfigTrendBU = NULL,
  ending = NULL), ...)
}
\arguments{
\item{librisk}{Data frame of the health risks}

\item{riskvar}{Variable name of the specific risk}

\item{risk.level}{The level of the risk. \cr
\describe{
 \item{0}{section title}
 \item{1}{risk group}
 \item{2}{specific risk}
}}

\item{title.headlevel}{If \code{risk.level} = 1, then \code{title.headlevel} should
be 2, and in markdown document, it is shown as "##"}

\item{lang}{Language, "cn" or "en"}

\item{suppl}{A list. Supplemental wording pre- or post- each element of the
section.}

\item{...}{}
}
\value{
A list containing all the wording elements in the list.
}
\description{
In an automated report, use this function to structure the elements you want
to put in each section.
}
\examples{
\dontrun{
getStrInterpret(libRisk, "Owt", suppl=list(posttitle="Sub title", ending="Close"))
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
