% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{histAge}
\alias{histAge}
\title{Histogram of Age + Gender}
\usage{
histAge(x = NULL, y, gender = NULL, label, parent = NULL,
  agegrp = NULL, ...)
}
\arguments{
\item{x}{BU or city vector}

\item{y}{Age vector}

\item{gender}{Gender vector}

\item{label}{A data frame to show in the parameter table}

\item{parent}{Parent GUI}

\item{agegrp}{Age grouping standard}

\item{...}{}
}
\value{
Nothing but a GUI showing histograms
}
\description{
Generate histogram of age + gender. It is used in \code{setRptParam()} function to
look into the age distribution.
}
\examples{
\dontrun{
histAge(y=Age, label=Label)
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
