% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstooltip.R
\name{JStooltip}
\alias{JStooltip}
\title{JavaScript Tooltip for Echarts}
\usage{
JStooltip(type, digits, ...)
}
\arguments{
\item{type}{The script type name. \cr
\itemize{
 \item axis_pct - Turn value to percents;\cr
 \item axis_abs_pct - Turn value to abs values and then percents;\cr
 \item tile_bar_pct - Turn value to percents by groups;\cr
 \item tile_bar_abs_pct - Turn value to abs values and then percents by groups;\cr
 \item bar_mkline_pct - Percentize the grouped values for bars and marklines;\cr
 \item mkline_pct - Percentize the grouped values for marklines
}}

\item{digits}{Number of digits, if applicable.}

\item{...}{ignore}
}
\value{
String. JS tooltip scripts.
}
\description{
Get specific JS tooltip scripts according to your needs. The scripts are used
for echarts.
}
\examples{
JStooltip("axis_pct", 1)
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
