% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParam.R
\name{queryAccess32}
\alias{queryAccess32}
\title{Get Data from 32-bit Access with 64-bit R}
\usage{
queryAccess32(db.path, pwd = NULL, qry.tbl = NULL, out.names = NULL,
  silent = FALSE)
}
\arguments{
\item{db.path}{charcter, MS Access db full path}

\item{pwd}{character, the password to the Access db if applicable. If there is
no password, set it NULL (default).}

\item{qry.tbl}{character vector, SQL queries or tables name.}

\item{out.names}{character vector, names of the output objects. Must be of the same
length as qry.tbl.}

\item{silent}{logical, if avoid any message. Deafult FALSE.}
}
\value{
An Rdata file (acc32.Rdata) in your work directory.
}
\description{
When x64 Access driver is missing, you cannot directly fetch data from MS Access 
using \pkg{RODBC} right now. This is a work-around function to initiate an 
independent x86 R session to fetch the data, and save the data.frames as an
\code{acc32.Rdata} file in your work directory.
}
\examples{
\dontrun{
queryAccess32("~/test.accdb", qry.tbl=c("User", "select * from Deal;"))
}

}
