% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_source.R
\name{restrImput}
\alias{restrImput}
\title{Restructure Mannually Imput Summarized Data}
\usage{
restrImput(imput, xvar, yvar, riskvar, librisk)
}
\arguments{
\item{imput}{Data frame. Summary input of checkup results}

\item{xvar}{Variable name to be crosstabed to x.}

\item{yvar}{Variable name to be crosstabed to y.}

\item{riskvar}{Risk variable name.}

\item{librisk}{Data frame of the risk pool.}
}
\value{
A restructured data frame.
}
\description{
Work with \code{\link{tabSource}}. The summarized data extracted from CheckupRawData
system is not working directly. You have to restructure it with this function.
}
\examples{

}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
