% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setParam.R
\name{setRptParam}
\alias{setRptParam}
\title{Set Report Parameters in GUI}
\usage{
setRptParam(db = NULL, ...)
}
\arguments{
\item{db}{the full path to the database. If Null, a wizard will guide you there.}

\item{...}{}
}
\value{
A list containing company, companyid, bu, buid, proj, projid,
agegrp, lang, curyr, dynamic, rankby
}
\description{
Set report parameters in a GUI wizard for automatic checkup reporting.
}
\examples{
\dontrun{
setRptParam()
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
