% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_age.R
\name{tabAge}
\alias{tabAge}
\title{Age-Gender Table in Auto Reporting}
\usage{
tabAge(data, imput = NULL, riskvar, librisk, lang = c("cn", "en"),
  sort = list(col = 1, desc = FALSE), digits = 1, dynamic = FALSE,
  ...)
}
\arguments{
\item{data}{Raw data frame.}

\item{imput}{Summarized data frame.}

\item{riskvar}{Variable name of the risk.}

\item{librisk}{Data frame of the risk pool.}

\item{lang}{'cn' or 'en'.}

\item{sort}{A list. Default \code{list(col=1, desc=FALSE)}, meaning sort the
results ascendingly by Col1.}

\item{digits}{Number of digits of the results. Default 1.}

\item{dynamic}{Interactive JavaScript charts/tables or not. Default FALSE.}

\item{...}{}
}
\value{
A \code{knitr} or \code{DT} object.
}
\description{
Only used for automated reporting. Get an age-gender distribution.
}
\examples{

}
\seealso{
\code{\link{knitr}}, \code{DT}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
