% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_bu.R
\name{tabBU}
\alias{tabBU}
\title{BU-spec Table in Auto Reporting}
\usage{
tabBU(data, imput = NULL, riskvar, librisk, tblBU, tblGender, tblCity,
  digits = 1, lang = c("cn", "en"), dynamic = FALSE, ...)
}
\arguments{
\item{data}{Raw data frame.}

\item{imput}{Summarized data frame.}

\item{riskvar}{Variable name of the risk.}

\item{librisk}{Data frame of the risk pool.}

\item{tblBU}{Data frame of BU pool.}

\item{tblGender}{Data frame of gender pool.}

\item{tblCity}{Data frame of city pool.}

\item{digits}{Number of digits of the results. Default 1.}

\item{lang}{'cn' or 'en'.}

\item{dynamic}{Interactive JavaScript charts/tables or not. If TRUE, call
\pkg{DT}::\code{\link{datatable}} to make the object. Default FALSE.}

\item{...}{}
}
\value{
A data frame.
}
\description{
Only used for automated reporting. Get a BU distribution.
}
\examples{

}
\seealso{
\code{\link{knitr}}  \code{\link{datatable}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
