% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demog_source.R
\name{tabDemogSource}
\alias{tabDemogSource}
\title{Demographic Data Source Table in Auto Reporting}
\usage{
tabDemogSource(data, imput = NULL, xvar, gvar, yvar = NULL,
  tblxvar = NULL, tblgvar = NULL, sort = list(col = 1, desc = FALSE),
  lang = LANG)
}
\arguments{
\item{data}{Raw data frame.}

\item{imput}{Summarized data frame.}

\item{xvar}{Variable name of x.}

\item{gvar}{Variable name of grouping factor.}

\item{yvar}{Variable name of y.}

\item{tblxvar}{Data frame of x variable mapping data.}

\item{tblgvar}{Data frame of g variable mapping data.}

\item{sort}{A list. Default \code{list(col=1, desc=FALSE)}, meaning sort the
resutls ascendingly by col1.}

\item{lang}{'cn' or 'en'.}
}
\value{
A data frame.
}
\description{
Only used for automated reporting. Get a demographic distribution.
}
\examples{

}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
