% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi.R
\name{tabEpiPrev}
\alias{tabEpiPrev}
\title{Epidemiologic Prevalence Table (data frame)}
\usage{
tabEpiPrev(data, imput = NULL, rowvar, colvar, riskvar, neg = "0",
  margins = NULL, lang = c("cn", "en"), cols = c("npos", "ntotal",
  "percent"), pct.digits = 0, colvar.together = FALSE,
  empty.cols = FALSE, na.string = .pkgPara()$na.string,
  sort = list(col = 1, desc = FALSE), col.levels = NULL, ...)
}
\arguments{
\item{data}{Raw dataset}

\item{imput}{Summarized data set}

\item{rowvar}{The variable used as rows}

\item{colvar}{The variable used as columns}

\item{riskvar}{The risk variable}

\item{neg}{Define which value represents negative. Default is 0.}

\item{margins}{Refer to margins argument in \code{reshape2::dcast}}

\item{lang}{language, "cn" or "en".}

\item{cols}{A vector. Which elements are to be included. Default is
\code{c("npos", "ntotal", "percent")}: number of positives, number of total, percent.}

\item{pct.digits}{Number of digits shown in percentages. If \code{pct.digits}==NULL,
then the function does not coerce prevalence rates to percents.}

\item{colvar.together}{Logical.}

\item{empty.cols}{Logical. If TRUE, empty columns will be added in the table
to separate columns sectors.}

\item{na.string}{A named vector \code{c(cn=..., en=...)}. How to present NAs
when \code{lang} is 'cn' or 'en'. Default \code{.pkgPara()$na.string}.}

\item{sort}{Sorting standard. If \code{sort}=list(col=1, desc=FALSE), then the
function sorts the data ascendingly by Col1. If \code{sort}=list(col=3,
desc=TRUE), the function sorts the data descendingly by Col3. You can also
directly assign a vector, e.g. list(col=1, val=1:4)}

\item{col.levels}{A vector re-defining the factor level of the variable to be
put as the column. Default null.}

\item{...}{}
}
\value{
A data frame restructured based on you requirements.
}
\description{
Draw an prevalence table with raw data combining with summarized data, in
addition with table controlling parameters.
}
\examples{

}
\seealso{
\code{\link{dcast}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
