% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_risk.R
\name{tabRisks}
\alias{tabRisks}
\title{Risk Aggregation Stats in Auto Reporting}
\usage{
tabRisks(data, imput = NULL, librisk, lang = c("cn", "en"), curYr,
  digits = 1, dynamic = FALSE, ...)
}
\arguments{
\item{data}{Raw data frame.}

\item{imput}{Summarized data frame.}

\item{librisk}{Data frame of the risk pool.}

\item{lang}{'cn' or 'en'.}

\item{curYr}{Integer. Current year used for reporting.}

\item{digits}{Number of digits of the results. Default 1.}

\item{dynamic}{Interactive JavaScript charts/tables or not. If TRUE, call
\pkg{DT}::\code{\link{datatable}} to make the object. Default FALSE.}

\item{...}{}
}
\value{
A data frame.
}
\description{
Only used for automated reporting. Get a summarized risk distribution.
}
\examples{

}
\seealso{
\code{\link{knitr}}  \code{\link{datatable}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
