% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_source.R
\name{tabSource}
\alias{tabSource}
\title{Data Frame for Tabulation in Auto Reporting}
\usage{
tabSource(data, imput = NULL, riskvar, xvar, yvar, xvar.imput = xvar,
  yvar.imput = yvar, tblRisk, tblBU, tblGender, tblCity, digits = 1,
  lang = c("cn", "en"), row_levels = NULL, col.levels = NULL,
  sort = list(col = 1, desc = FALSE),
  na.string = getOption("na.string"), ...)
}
\arguments{
\item{data}{Raw dataset}

\item{imput}{Summarized data set}

\item{riskvar}{The risk variable}

\item{xvar}{Variable name of x in raw data set.}

\item{yvar}{Variable name of y in raw data set.}

\item{xvar.imput}{Variable name of x in summarized data set.}

\item{yvar.imput}{Variable name of y in summarized data set.}

\item{tblRisk}{Data frame of risk pool.}

\item{tblBU}{Data frame of BU pool.}

\item{tblGender}{Data frame of gender pool.}

\item{tblCity}{Data frame of city pool.}

\item{digits}{Number of digits in results.}

\item{lang}{Language, 'cn' or 'en'.}

\item{row_levels}{A vector re-defining the factor level of the variable to be
put as the raw. Default null. If 'asis', keep the original order; if 'NULL',
reorder the columns; if a specific vector, sort columns by that order.}

\item{col.levels}{A vector re-defining the factor level of the variable to be
put as the column. Default null.}

\item{sort}{Sorting standard. If \code{sort}=list(col=1, desc=FALSE), then the
function sorts the data ascendingly by Col1. If \code{sort}=list(col=3,
desc=TRUE), the function sorts the data descendingly by Col3. You can also
directly assign a vector, e.g. \code{list(col=1, val=1:4)}.}

\item{na.string}{A named vector \code{c(cn=..., en=...)}. How to present NAs
when \code{lang} is EN or CN. Default \code{getOption("na.string")}.}

\item{...}{ignore}
}
\value{
A data frame restructured based on you requirements.
}
\description{
Draw a prevalence table with raw data combining with summarized data, in
addition with table controlling parameters. The parameter pool data frames
are yielded using \code{setRptParam()}.
}
\details{
It is more universal than \code{tabEpiPrev()}.
}
\examples{


}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
