% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bu_trend.R
\name{tabTrendBU}
\alias{tabTrendBU}
\title{BU Trend Table in Auto Reporting}
\usage{
tabTrendBU(data, imput = NULL, riskvar, librisk, tblBU, tblGender,
  sort = list(col = 1, desc = FALSE), digits = 1, lang = c("cn",
  "en"), dynamic = FALSE, ...)
}
\arguments{
\item{data}{Raw data frame.}

\item{imput}{Summarized data frame.}

\item{riskvar}{Variable name of the risk.}

\item{librisk}{Data frame of the risk pool.}

\item{tblBU}{Data frame of the BU pool.}

\item{tblGender}{Data frame of gender pool.}

\item{sort}{A list. Default \code{list(col=1, desc=FALSE)}, meaning sort the
results ascendingly by Col1.}

\item{digits}{Number of digits of the results. Default 1.}

\item{lang}{'cn' or 'en'.}

\item{dynamic}{Interactive JavaScript charts/tables or not. If TRUE, call
\pkg{DT}::\code{\link{datatable}} to make the object. Default FALSE.}

\item{...}{}
}
\value{
A \code{knitr} or \code{DT} object.
}
\description{
Only used for automated reporting. Get an age-gender distribution.
}
\examples{

}
\seealso{
\code{\link{knitr}}  \code{\link{datatable}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
